# Concatenate restructure & outer join for layers