This is a refactor of the `concatenate` method to make it easier to follow and extend, as well as to add some new features.

* Fixes #351 by allowing for concatenation of layers
* Gets us closer to closing #295
* Dataframes in `obsm` keep being dataframes after concatenation
* Categorical columns can keep being categorical after concatenation (basically, pandas rules get followed)

This moves pretty much all of the concatenation logic to the `anndata._core.merge` module, in particular the `concat` function. At some point I would like to deprecate `AnnData.concatenate` and replace it with a `ad.concat`/ `sc.concat` function. There is more work to do before that (most importantly, hammering out the API), but this is a good step forward.

## Breaking change

This was reverted.

~~Currently, there is only one, but I think it's more of a fix really. Previously, an outer join would fill missing values in numpy arrays with `nan`, while sparse arrays would use zeros. This kinda breaks most of what we've been trying to with pretending numpy arrays and scipy sparse arrays are interchangeable.~~

## Performance

This makes outer concatenation significantly faster.

Using this code:

```python
import numpy as np
from anndata.tests.helpers import gen_adata

def permuted_var_adata(nobs, nvar, nvar_kept): 
    adata = gen_adata((nobs, nvar)) 
    return adata[:, np.random.choice(adata.var_names, nvar_kept, replace=False)].copy() 

adatas = [permuted_var_adata(1000, 1000, 900) for i in range(10)]


%timeit adatas[0].concatenate(adatas[1:])
%timeit adatas[0].concatenate(adatas[1:], join="outer")
```

|                          |   inner join |   outer join |
|:-------------------------|-------------:|-------------:|
| master (9e0767b761)      |          240 ms ± 20.5 ms |          1.6 s ± 16 ms |
| current HEAD (0b14e1610) |          237 ms ± 12.4 ms |          304 ms ± 11.2 ms |

## TODO

- [ ] This could probably use a more thorough set of tests. 
- [x] I would also like to deal with the `obsm` issue by making the `join` keyword apply to those entries.
- [x] Black really doesn't like fluent style when it's only two lines. I should make that pretty again.
- [ ] Doc string for concatenate needs to be updated