@giovp 2020-04-14
Hi @ivirshup , just saw this, really cool! From my understanding, this should also close #338 , as `join='outer'` would return `obsm` with missing values for obs entries that did not have that `obsm` in the original data. Does it make sense?