@ivirshup 2020-04-15
@giovp, yes. I wasn't sure it made sense to add another argument for this. To me concatenating the elements of `obsm` seems a lot like concatenating `X` or elements of `layers`, so I figured this could keep the solution simple.