@falexwolf 2020-04-23
This is great, Isaac! Only remaining Q the fill value discussion on Slack. Otherwise happy if you merge!