@ivirshup 2020-04-29
I believe this is pretty much ready to go.

Notes on other things that should get done in future

* The code for `fill_values` is a bit of a mess, and should be cleaned up at some point. It's just passed around too much.
* In the recent commits I also realized I wasn't performing a full outer join on the layers ( layers that were only present in one object wouldn't show up in the final object), so that's been fixed, but should get a test. I can add this separately so @giovp can get going.