@gheimberg 2020-05-08
Thanks for this awesome improvement @ivirshup ! Trying this out now and I sometimes get a 

```pytb
---------------------------------------------------------------------------
ValueError                                Traceback (most recent call last)
<timed eval> in <module>

~/.anaconda3/envs/dev/lib/python3.7/site-packages/anndata/_core/anndata.py in concatenate(self, join, batch_key, batch_categories, uns_merge, index_unique, fill_value, *adatas)
   1691             uns_merge=uns_merge,
   1692             fill_value=fill_value,
-> 1693             index_unique=index_unique,
   1694         )
   1695 

~/.anaconda3/envs/dev/lib/python3.7/site-packages/anndata/_core/merge.py in concat(adatas, join, batch_key, batch_categories, uns_merge, index_unique, fill_value)
    472     elif join == "outer":
    473         layers = outer_concat_aligned_mapping(
--> 474             [a.layers for a in adatas], reindexers, fill_value=fill_value
    475         )
    476         obsm = outer_concat_aligned_mapping(

~/.anaconda3/envs/dev/lib/python3.7/site-packages/anndata/_core/merge.py in outer_concat_aligned_mapping(mappings, reindexers, index, fill_value)
    389             cur_reindexers,
    390             index=index,
--> 391             fill_value=fill_value,
    392         )
    393     return result

~/.anaconda3/envs/dev/lib/python3.7/site-packages/anndata/_core/merge.py in concat_arrays(arrays, reindexers, index, fill_value)
    309             [
    310                 f(as_sparse(a), fill_value=fill_value)
--> 311                 for f, a in zip(reindexers, arrays)
    312             ],
    313             format="csr",

~/.anaconda3/envs/dev/lib/python3.7/site-packages/anndata/_core/merge.py in <listcomp>(.0)
    309             [
    310                 f(as_sparse(a), fill_value=fill_value)
--> 311                 for f, a in zip(reindexers, arrays)
    312             ],
    313             format="csr",

~/.anaconda3/envs/dev/lib/python3.7/site-packages/anndata/_core/merge.py in reindexer(X, fill_value)
    274             dtype=out_dtype,
    275         )
--> 276         out = X @ idxmtx
    277 
    278         if fill_value != 0:

~/.anaconda3/envs/dev/lib/python3.7/site-packages/scipy/sparse/base.py in __matmul__(self, other)
    562             raise ValueError("Scalar operands are not allowed, "
    563                              "use '*' instead")
--> 564         return self.__mul__(other)
    565 
    566     def __rmatmul__(self, other):

~/.anaconda3/envs/dev/lib/python3.7/site-packages/scipy/sparse/base.py in __mul__(self, other)
    481         if issparse(other):
    482             if self.shape[1] != other.shape[0]:
--> 483                 raise ValueError('dimension mismatch')
    484             return self._mul_sparse_matrix(other)
    485 

ValueError: dimension mismatch
```

@gokceneraslan suggested I look at what `.layers` and saw that one anndata has a counts layer while the other did not. Would be nice to be able to concatenate even when layers don't match up perfectly.